Se edito eliminar producto totalmente

+++ js/services/eliminarProducto.js
@@ -1,22 +1,18 @@
 import { ConexionApi } from "../controllers/conexionAPI.js";
 
-
-document.addEventListener('DOMContentLoaded', () =>{
-    requestAnimationFrame(() => {
-        const botones = document.querySelectorAll('[data-id]');
-        
-        botones.forEach( (boton) => {
-            boton.addEventListener("click",  () => {
-                const id = boton.attributes[1].value;
-                // console.log(`id ${id}`);
-
-                 ConexionApi.eliminarArticulo(id);
-
-            })
-        } )
-
-    });
-});
-
+export async function handleDelete(event) {
+    const id = event.target.getAttribute('data-id');
+    const confirmacion = confirm("¿Deseas eliminar este artículo?");
+    
+    if (confirmacion) {
+       try {
+          await ConexionApi.eliminarArticulo(id);
+          event.target.closest('.card').remove();
+       } catch (error) {
+          console.error("Error al eliminar el producto:", error);
+          alert("Hubo un problema al eliminar el producto. Inténtalo de nuevo más tarde.");
+       }
+    }
+ }
